+++ R/gridinfo.R
@@ -2,8 +2,9 @@
 #'
 #' @description Return a list containing information of a regular grid / domain
 #'
-#' @param file file name/path to a wrfinput of wrfchemi file
+#' @param file file name/path to a wrfinput, wrfchemi or geog_em file
 #' @param z TRUE for read wrfinput vertical coordinades
+#' @param geo True for use geog_em files
 #' @param verbose display additional information
 #'
 #' @note just WRF-Chem is suported by now
@@ -31,31 +32,44 @@
 #' text(grid_d3$xlim[1],grid_d3$Ylim[2],"d3",pos=2, offset = 0.0)
 #'}
 
-gridInfo <- function(file = file.choose(),z=F,verbose = T){
+gridInfo <- function(file = file.choose(),z=F,geo = F,verbose = T){
     if(verbose)
       cat(paste("Grid information from:",file,"\n"))
     wrf <- ncdf4::nc_open(file)
-    lat <- ncdf4::ncvar_get(wrf,varid = "XLAT")
-    lon <- ncdf4::ncvar_get(wrf,varid = "XLONG")
+    if(geo){
+      lat <- ncdf4::ncvar_get(wrf,varid = "XLAT_M")      # nocov
+      lon <- ncdf4::ncvar_get(wrf,varid = "XLONG_M")     # nocov
+    }else{
+      lat <- ncdf4::ncvar_get(wrf,varid = "XLAT")
+      lon <- ncdf4::ncvar_get(wrf,varid = "XLONG")
+    }
+
     time<- ncdf4::ncvar_get(wrf,varid = "Times")
-    dx  <- ncdf4::ncatt_get(wrf,varid = 0,attname = "DX")$value / 1000 # km
+    dx  <- ncdf4::ncatt_get(wrf,varid = 0,attname = "DX")$value / 1000 # to km
     if(z){
-      PHB <- ncdf4::ncvar_get(wrf,varid = "PHB") # 3d
-      PH  <- ncdf4::ncvar_get(wrf,varid = "PH")  # 3d
-      HGT <- ncdf4::ncvar_get(wrf,varid = "HGT") # 2d
+      PHB <- ncdf4::ncvar_get(wrf,varid = "PHB")        # 3d
+      PH  <- ncdf4::ncvar_get(wrf,varid = "PH")         # 3d
+      HGT <- ncdf4::ncvar_get(wrf,varid = "HGT")        # 2d
       z   <- PH
-      for(i in 1:dim(PH)[3]){
-        z[,,i]   <- (PH[,,i] + PHB[,,i])/9.8 - HGT # 9.81 return values < 0, ~10-5
+      if(length(time) == 1){                            # just one time
+        for(i in 1:dim(PH)[3]){
+          z[,,i]   <- (PH[,,i] + PHB[,,i])/9.8 - HGT    # 9.81 return values < 0, ~10-5
+        }                                               # this is for an alternative use
+      }else{                                            # for multiple times (test version)
+        for(i in 1:dim(PH)[3]){                         # nocov
+          z[,,i,]   <- (PH[,,i,] + PHB[,,i,])/9.8 - HGT # nocov
+        }
       }
+
     }else{
       z <- NA
     }
     ncdf4::nc_close(wrf)
     lx  <- range(lon)
     ly  <- range(lat)
-    OUT <- list(Times = time, Lat = lat, Lon = lon, Horizontal = dim(lat),
-                Levels = z, DX = dx,xlim = lx, Ylim = ly, File = file,
+    OUT <- list(File = file, Times = time, Lat = lat, Lon = lon, z = z,
+                Horizontal = dim(lat), DX = dx, xlim = lx, Ylim = ly,
                 Box = list(x = c(lx[2],lx[1],lx[1],lx[2],lx[2]),
-                           y = c(ly[2],ly[2],ly[1],ly[1],ly[2])),z = z)
+                           y = c(ly[2],ly[2],ly[1],ly[1],ly[2])))
     return(OUT)
 }